+++ system/lib/pthread/library_pthread.c
@@ -65,54 +65,32 @@ int sched_get_priority_min(int policy) {
     return 0;
 }
 
-int pthread_mutexattr_getprioceiling(const pthread_mutexattr_t *restrict attr, int *restrict prioceiling)
-{
+int pthread_mutexattr_getprioceiling(const pthread_mutexattr_t *restrict attr, int *restrict prioceiling) {
   // Not supported either in Emscripten or musl, return a faked value.
   if (prioceiling) *prioceiling = 99;
   return 0;
 }
 
-int pthread_mutexattr_setprioceiling(pthread_mutexattr_t *attr, int prioceiling)
-{
+int pthread_mutexattr_setprioceiling(pthread_mutexattr_t *attr, int prioceiling) {
   // Not supported either in Emscripten or musl, return an error.
   return EPERM;
 }
 
-static uint32_t dummyZeroAddress = 0;
-
 void emscripten_thread_sleep(double msecs) {
-  double now = emscripten_get_now();
-  double target = now + msecs;
-
-  // If we have less than this many msecs left to wait, busy spin that instead.
-  double min_ms_slice_to_sleep = 0.1;
-
-  // Break up sleeping so that we process proxied work at regular intervals.
-  // TODO(sbc): This should be removed and/or moved down into
-  // `emscripten_futex_wait`.
-  double max_ms_slice_to_sleep = 100;
-
-  emscripten_conditional_set_current_thread_status(
-    EM_THREAD_STATUS_RUNNING, EM_THREAD_STATUS_SLEEPING);
-
-  do {
-    // Keep processing the main loop of the calling thread.
-    __pthread_testcancel(); // pthreads spec: sleep is a cancellation point, so must test if this
-                            // thread is cancelled during the sleep.
-    emscripten_current_thread_process_queued_calls();
-
-    now = emscripten_get_now();
-    double ms_to_sleep = target - now;
-    if (ms_to_sleep < min_ms_slice_to_sleep)
-      continue;
-    if (ms_to_sleep > max_ms_slice_to_sleep)
-      ms_to_sleep = max_ms_slice_to_sleep;
-    emscripten_futex_wait(&dummyZeroAddress, 0, ms_to_sleep);
-    now = emscripten_get_now();
-  } while (now < target);
-
-  emscripten_conditional_set_current_thread_status(
-    EM_THREAD_STATUS_SLEEPING, EM_THREAD_STATUS_RUNNING);
+  // We include emscripten_current_thread_process_queued_calls before and
+  // after sleeping since that is how we recieve "async" signals.
+  // We include __pthread_testcancel here becuase clock_nanosleep is
+  // a pthread cancelation point.
+  emscripten_current_thread_process_queued_calls();
+  __pthread_testcancel();
+  emscripten_conditional_set_current_thread_status(EM_THREAD_STATUS_RUNNING,
+                                                   EM_THREAD_STATUS_SLEEPING);
+  uint32_t dummyZeroAddress = 0;
+  emscripten_futex_wait(&dummyZeroAddress, 0, msecs);
+  emscripten_conditional_set_current_thread_status(EM_THREAD_STATUS_SLEEPING,
+                                                   EM_THREAD_STATUS_RUNNING);
+  emscripten_current_thread_process_queued_calls();
+  __pthread_testcancel();
 }
 
 static struct pthread __main_pthread;