+++ test/test_core.py
@@ -138,7 +138,6 @@ def wasm_simd(func):
 
   @wraps(func)
   def decorated(self, *args, **kwargs):
-    self.require_simd()
     if self.is_wasm2js():
       self.skipTest('wasm2js only supports MVP for now')
     if '-O3' in self.cflags:
@@ -156,6 +155,18 @@ def asan(func):
   @no_wasm2js('TODO: ASAN in wasm2js')
   @no_wasm64('TODO: ASAN in memory64')
   @no_2gb("asan doesn't support GLOBAL_BASE")
+  def decorated(self, *args, **kwargs):
+    return func(self, *args, **kwargs)
+
+  return decorated
+
+
+def requires_wasm_workers(func):
+  assert callable(func)
+
+  @wraps(func)
+  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
+  @no_sanitize('sanitizers do not support WASM_WORKERS')
   def decorated(self, *args, **kwargs):
     return func(self, *args, **kwargs)
 
@@ -204,7 +215,7 @@ def requires_x64_cpu(func):
 
   @wraps(func)
   def decorated(self, *args, **kwargs):
-    if platform.machine().lower() not in ['x86_64', 'amd64']:
+    if platform.machine().lower() not in {'x86_64', 'amd64'}:
       return self.skipTest(f'This test requires a native x64 CPU. Current CPU is {platform.machine()}.')
     return func(self, *args, **kwargs)
 
@@ -288,13 +299,11 @@ def also_with_wasm_workers(func):
 
   @wraps(func)
   def metafunc(self, ww, *args, **kwargs):
+    f = func
     if ww:
-      if self.get_setting('WASM_ESM_INTEGRATION'):
-        self.skipTest('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
-      if is_sanitizing(self.cflags):
-        self.skipTest('sanitizers are not compatible with WASM_WORKERS')
+      f = requires_wasm_workers(f)
       self.cflags += ['-sWASM_WORKERS']
-    return func(self, *args, **kwargs)
+    return f(self, *args, **kwargs)
 
   parameterize(metafunc, {'': (False,), 'ww': (True,)})
   return metafunc
@@ -992,8 +1001,7 @@ class TestCoreBase(RunnerCore):
   def test_longjmp(self):
     self.do_core_test('test_longjmp.c')
 
-  @no_sanitize('sanitizers do not support WASM_WORKERS')
-  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
+  @requires_wasm_workers
   def test_longjmp_wasm_workers(self):
     self.do_core_test('test_longjmp.c', cflags=['-sWASM_WORKERS'])
 
@@ -3370,7 +3378,7 @@ Var: 42
     data_exports = get_data_exports('test_dlfcn_self.wasm')
     # Certain exports are removed by wasm-emscripten-finalize, but this
     # tool is not run in all configurations, so ignore these exports.
-    data_exports = [d for d in data_exports if d not in ('__start_em_asm', '__stop_em_asm')]
+    data_exports = [d for d in data_exports if d not in {'__start_em_asm', '__stop_em_asm'}]
     data_exports = '\n'.join(sorted(data_exports)) + '\n'
     self.assertFileContents(test_file('core/test_dlfcn_self.exports'), data_exports)
 
@@ -4025,9 +4033,7 @@ caught outer int: 123
     if getattr(self, 'dylink_reversed', False):
       # Test the reverse case.  There we flip the role of the side module and main module.
       # - We add --no-entry since the side module doesn't have a `main`
-      side_ = side
-      side = main
-      main = side_
+      side, main = main, side
     self.maybe_closure()
     # Same as dylink_test but takes source code as filenames on disc.
     old_args = self.cflags.copy()
@@ -7773,8 +7779,7 @@ void* operator new(size_t size) {
     '''
     self.do_run(src, '418\ndotest returned: 42\n', cflags=['-lembind', '-fno-rtti', '-frtti'])
 
-  @no_sanitize('sanitizers do not support WASM_WORKERS')
-  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
+  @requires_wasm_workers
   def test_embind_wasm_workers(self):
     self.do_run_in_out_file_test('embind/test_embind_wasm_workers.cpp', cflags=['-lembind', '-sWASM_WORKERS'])
 
@@ -7901,7 +7906,7 @@ void* operator new(size_t size) {
     # optimizer can deal with both types.
     map_filename = map_referent + '.map'
 
-    data = json.load(open(map_filename))
+    data = json.loads(utils.read_file(map_filename))
     if hasattr(data, 'file'):
       # the file attribute is optional, but if it is present it needs to refer
       # the output file.
@@ -8193,7 +8198,7 @@ void* operator new(size_t size) {
   # Test that a main with arguments is automatically asyncified.
   @with_asyncify_and_jspi
   def test_async_main(self):
-    create_file('main.c',  r'''
+    create_file('main.c', r'''
 #include <stdio.h>
 #include <emscripten.h>
 int main(int argc, char **argv) {
@@ -8209,7 +8214,7 @@ int main(int argc, char **argv) {
     # needs to flush stdio streams
     self.set_setting('EXIT_RUNTIME')
 
-    create_file('main.c',  r'''
+    create_file('main.c', r'''
 #include <stdio.h>
 #include <emscripten.h>
 void f(void *p) {
@@ -8230,7 +8235,7 @@ int main() {
 
   @with_asyncify_and_jspi
   def test_async_loop(self):
-    create_file('main.c',  r'''
+    create_file('main.c', r'''
 #include <stdio.h>
 #include <emscripten.h>
 int main() {
@@ -9253,17 +9258,16 @@ NODEFS is no longer included by default; build with -lnodefs.js
     self.do_core_test('pthread/emscripten_atomics.c', cflags=['-pthread'])
 
   @requires_pthreads
-  def test_emscripten_futexes(self):
+  def test_emscripten_futex_api_basics(self):
     # This test explicitly checks behavior of passing NULL to emscripten_futex_wake() so
     # need to disable the `-Wno-nonnull` to disabled these warnings.
-    self.do_core_test('pthread/emscripten_futexes.c', cflags=['-pthread', '-Wno-nonnull'])
+    self.do_core_test('pthread/emscripten_futex_api_basics.c', cflags=['-pthread', '-Wno-nonnull'])
 
   @requires_pthreads
   def test_stdio_locking(self):
     self.do_core_test('test_stdio_locking.c', cflags=['-sPTHREAD_POOL_SIZE=2'])
 
-  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
-  @no_sanitize('sanitizers do not support WASM_WORKERS')
+  @requires_wasm_workers
   def test_stdio_locking_ww(self):
     # Note: do not combine with test_stdio_locking above because we want to test standalone
     # wasm workers here and `@requires_pthreads` would prevent that.
@@ -9700,33 +9704,32 @@ NODEFS is no longer included by default; build with -lnodefs.js
     self.run_process([FILE_PACKAGER, 'test.data', '--preload', 'file1.txt', 'file2.txt', '--from-emcc', '--js-output=script2.js'])
     self.do_runf('test_emscripten_async_load_script.c', cflags=['-sFORCE_FILESYSTEM'])
 
-  @no_sanitize('sanitizers do not support WASM_WORKERS')
   @also_with_minimal_runtime
   @also_with_modularize
-  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
+  @requires_wasm_workers
   def test_wasm_worker_hello(self):
     if self.is_wasm2js() and '-sMODULARIZE' in self.cflags:
       self.skipTest('WASM2JS + MODULARIZE + WASM_WORKERS is not supported')
     self.maybe_closure()
     self.do_run_in_out_file_test('wasm_worker/hello_wasm_worker.c', cflags=['-sWASM_WORKERS'])
 
-  @no_sanitize('sanitizers do not support WASM_WORKERS')
-  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
+  @requires_wasm_workers
+  def test_wasm_worker_exceptions(self):
+    self.do_runf('wasm_worker/test_wasm_worker_exceptions.c', 'worker sent an error! Aborted', assert_returncode=NON_ZERO, cflags=['-sWASM_WORKERS'])
+
+  @requires_wasm_workers
   def test_wasm_worker_malloc(self):
     self.do_run_in_out_file_test('wasm_worker/malloc_wasm_worker.c', cflags=['-sWASM_WORKERS'])
 
-  @no_sanitize('sanitizers do not support WASM_WORKERS')
-  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
+  @requires_wasm_workers
   def test_wasm_worker_runtime_debug(self):
     self.do_runf('wasm_worker/hello_wasm_worker.c', 'wasm worker starting ...', cflags=['-sWASM_WORKERS', '-sRUNTIME_DEBUG'])
 
-  @no_sanitize('sanitizers do not support WASM_WORKERS')
-  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
+  @requires_wasm_workers
   def test_wasm_worker_futex_wait(self):
     self.do_runf('wasm_worker/wasm_worker_futex_wait.c', cflags=['-sWASM_WORKERS'])
 
-  @no_sanitize('sanitizers do not support WASM_WORKERS')
-  @no_esm_integration('WASM_ESM_INTEGRATION is not compatible with WASM_WORKERS')
+  @requires_wasm_workers
   def test_wasm_worker_wait_async(self):
     self.do_runf('atomic/test_wait_async.c', cflags=['-sWASM_WORKERS'])